Changes the Location object to double from float

+++ Test/Backend/src/Location.h
@@ -5,11 +5,11 @@
 using namespace std;
 
 struct Location {
-	float longitude = 0;
-	float latitude = 0;
+	double longitude = 0;
+	double latitude = 0;
 	string address;
 
 	Location() {};
-	Location(string address, float longitude, float latitude) : address(address), longitude(longitude), latitude(latitude) {};
+	Location(string address, double longitude, double latitude) : address(address), longitude(longitude), latitude(latitude) {};
 	
 };
